:art: Imporve code structure for testing
