+++ .eslintrc
@@ -24,6 +24,7 @@
     "react/destructuring-assignment": ["off"],
     "react/jsx-props-no-spreading": ["off"],
     "no-console": ["error"],
-    "no-debugger": ["error"]
+    "no-debugger": ["error"],
+    "import/prefer-default-export": ["off"]
   }
 }
