+++ helpers/api/queries/locations.js
@@ -1,8 +1,6 @@
-import { gql, useQuery } from '@apollo/client'
-import ItemsList from '../components/Items/ItemsList'
+import { gql } from '@apollo/client'
 
-const RYCK_AND_MORTY_DATA = gql`
-  # Write your query or mutation here
+export const RYCK_AND_MORTY_DATA = gql`
   query getRickAndMortyData($name: String!, $dimension: String!) {
     locations(filter: { name: $name, dimension: $dimension }) {
       results {
@@ -34,20 +32,3 @@ const RYCK_AND_MORTY_DATA = gql`
     }
   }
 `
-
-const Items = () => {
-  const { loading, error, data } = useQuery(RYCK_AND_MORTY_DATA, {
-    variables: {
-      name: 'Earth',
-      dimension: 'Replacement Dimension',
-    },
-  })
-
-  if (loading) return <p>Loading...</p>
-  if (error) return <p>Error :(</p>
-
-  return <ItemsList residents={data.locations.results[0].residents} />
-}
-
-export { RYCK_AND_MORTY_DATA }
-export default Items
