+++ pages/index.js
@@ -1,12 +1,12 @@
 import { getDataFromTree } from '@apollo/react-ssr'
 import withApollo from '../lib/withApollo'
-import Layout from '../components/Layout'
-import Items from '../containers/Items'
+import Layout from '../components/Layout/Layout'
+import List from '../components/List/List'
 
 const Homepage = () => {
   return (
     <Layout>
-      <Items />
+      <List />
     </Layout>
   )
 }
